Dockerfile: Use host architecture during builder stage.
This commit uses host architecture during builder stage but generated binaries
are built for the target architecture.
Due to use of CGO, we need a cross compiler and libraries for the target
architecture.
This is particularly useful when cross-building as it will avoid this stage to
be emulated.
As a result, this commit reduces building time of the default image for arm64
from 35 minutes to 15.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>