pkg/gadgets: Move code related to Kubernetes operator to different package
Code in pkg/gadgets is composed by:
- tracer/core: gadget tracer implementation using CO-RE
- tracer/standard: gadget tracer implementation on top of BCC
- gadget.go: glue code that implements the TraceFactory interface

The code in gadget.go is not related to the tracers itself but it's
used to "glue" the trace controller and the trace implementation. In
order to improve the code organization (by providing a better separation
of concerns) this commit moves all the gadget.go files to
pkg/gadget-collection.

There is not any change in functinality introduced by this commit, all
changes are files movements and renaming of import paths.