trace/network: fix race on containerDetach
The trace/network gadget checks the content of a global BPF map every
second in a go routine. When a pod is deleted, the BPF program for that
network namespace is removed in a different go routine.

This patch handles the pod deletion in the same go routine and make sure
that the BPF map is read before.

Signed-off-by: Alban Crequy <albancrequy@microsoft.com>