cmd/kubectl-gadget: Do not fail if tracer controller is dead on node
Add logic to handle the case when some the trace controller is dead
in some nodes, i.e. some traces don't reach the expected state.

This is important because it should be possible to get some results
from the cluster even if the gadget pod is not running fine in all of
the nodes.