cmd/kubect-gadget: Add readiness probe
2a233cad91dc ("cmd/kubectl-gadget: Use k8s API for deploy operation.")
introduced some logic in the deploy command to wait until the gadget
pods are ready to process operations. The logic checks that the number
of ready pods in the daemon set is the desired one.

Even if the logic was right, (it waited until all pods where Ready),
there was a missing piece. It could happen that the gadget tracer
manager is started but it's not ready to take requests yet becuase its
initialization isn't done. This issue happens sometimes when we try to
run a gadget just after deploying Inspektor Gadget on a heavy loaded
machine:

```
$ ./kubectl-gadget deploy && ./kubectl-gadget snapshot process
Creating Namespace/gadget...
Creating ServiceAccount/gadget...
Creating Role/gadget-role...
Creating RoleBinding/gadget-role-binding...
Creating ClusterRole/gadget-cluster-role...
Creating ClusterRoleBinding/gadget-cluster-role-binding...
Creating DaemonSet/gadget...
Creating CustomResourceDefinition/traces.gadget.kinvolk.io...
Waiting for gadget pod(s) to be ready...
0/2 gadget pod(s) ready
1/2 gadget pod(s) ready
2/2 gadget pod(s) ready
Error: timed out waiting for the condition
```

This commit adds readiness probes to fix this issue, the pods are only
marked as "Ready" when the gadget tracer manager is ready to process
requests.