standardtracers/trace/bind: Fix json unmarshalling error
The json produced by the bindsnoop BCC tools uses 0 (integer) for
representing the iface. Inspektor Gadget uses a string hence it was
failing with `json: cannot unmarshal number into Go struct field Event.if of type string`.

This commit introduces logic to parse that field as an integer and
use netlink to get the name of the interface.