ci: Remove references to development container repository.
Commit d49184649ec3 ("ci: switch to artifact based minikube testing") made use
of artifact to push the image to Minikube to run our integration tests.
This same commit modified the CI to run integration tests on ARO only on push
to main and tags.

As a consequence, the development container repository is no more used, so we
can totally remove reference to it in CI code.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>