cm/kubectl-gadget: Create traces based on gadget pods
The current logic creates traces on the existing nodes. The next commit
will add a node selector to the gadget daemonset hence this logic is
updated to create traces according to the gadget pods.

This commit also adds some more error messages in cases the user tries
to run it without deploying IG or in a node where the gadget pod is not
running.