btfgen: Check BTF file exists before calling btfgen.
A recent addition in btfhub-archive added an archive which did not contain a
BTF file [1].
As a consequence, our Makefile was failing because it expects archive to contain
BTF file, this leaded to failures in our build system.

So, this commit checks if the BTF files exists before calling bpftool.
If the file does not exists a message is printed on the standard error.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1] https://github.com/aquasecurity/btfhub-archive/pull/6