cli: Don't use BaseParser to print events of trace category
This commit is an intermediate step needed to progressively introduce
the usage of the columns library. Otherwise, we would need to update all
the trace gadgets at once, and it's what we are trying to avoid.

Once all the trace gadgets use the columns library, we will move the
logic to print JSON or columns to a common place (parser-tableformatter.go).