liveness: Rely on probe.timeoutSeconds but keeping a fallback
Currently, the probes will time out with the minimum value between
{livenessProbe|readinessProbe}.timeoutSeconds and our internal
clientTimeout. In fact, changing timeoutSeconds to any value larger
than our 2 seconds doesn't have any effect, as 2 seconds is the internal
timeout (clientTimeout) for the client that is performing the probe.

With this commit, the idea is start relying on probe.timeoutSeconds,
(which can be easily modified for testing) but keep using our internal
client's timeout just as fallback, when the timeoutSeconds feature
doesn't work, see
https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#configure-probes.