container-collection: Lazily enrich ownerReferences
Owner references are used by the secccomp advisor gadget to include
this information in the generated profiles. Getting this information
takes up to hundreds of milliseconds according to my experiments.
This makes the start of the gadget tracer manager slow in nodes with
many container.  This information is needed in few cases
(seccomp advisor), hence we can change the approach to only fill it
when needed.