advise/seccomp: Fix logic around container termination notification
It's possible to create multiple instances of the seccomp advisor gadget
, however it internally uses a single eBPF program that captures all
the syscalls on the host.

The current logic is wrong as it only subscribes to the container
termination events taking into consideration the filter spec of the
first seccomp advisor instance. Hence, the following case doesn't work:

```
$ kubectl run mypod1 --restart=Never --image=praqma/network-multitool \
	-- bash -c "while true; do ping -c 1 8.8.8.8 && sleep 1; done"
$ kubectl run mypod2 --restart=Never --image=praqma/network-multitool \
	-- bash -c "while true; do ping -c 1 8.8.8.8 && sleep 1; done"

$ kubectl -n gadget get seccompprofile
No resources found in gadget namespace.

$ ./kubectl-gadget advise seccomp-profile start -m seccomp-profile -p mypod1

$ ./kubectl-gadget advise seccomp-profile start -m seccomp-profile -p mypod2

$ kubectl delete pod mypod2

$ kubectl -n gadget get seccompprofile
No resources found in gadget namespace
```

After this commit, a profile is created for mypod2 by following the
same steps.