Fix runcfanotify dropping events for docker-ce on WSL2 / Ubuntu 22.04
On Ubuntu 22.04 WSL2, /var/run is symlinked to /run.
Docker-CE generates a containerd configuration that sets the state path
to /var/run/docker/containerd/daemon, which containerd uses to construct
runc's --pid-file path. However, the fanotify event path uses /run,
and since the paths don't match, runcfanotify returns false.

This had the effect of "dropping" the container create event,
so that (1) the container did not show up in trace results, and
(2) on SIGINT, the runcfanotify Close() method would hang waiting on
the read sycall in `pidFileDirNotify.GetEvent()`.

Fix it by using `os.Samefile` to compare the device/inode of the
two files instead of comparing paths.

Resolves: https://github.com/kinvolk/inspektor-gadget/issues/958