integration: Do not check args for trace exec on arm64.
There was a kernel bug which impedes tracing execve family syscall [1].
This bug was backported to all LTS stables.

But to permit running trace exec on other kernel, we decided to use kprobes
instead of tracepoint when running on arm64.
This was added in
commit 243759db6b19 ("pkg/gadgets: Use kprobe for execsnoop on arm64.").
The drawback is that kprobes cannot trace syscall arguments.

So, we disable this check to have the test pass on arm64.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=de6921856f99