integration: Use ; instead of && in script
We want to run the sleep regardless of the result of the previous
command, hence we have to use ;.
When && is used and the command fails, it keeps trying, what causes
to generate a lot of events, then making it very difficult to handle
the logs. (I saw a 180k log file from the trace capabilities gadget).