cmd: Only use sigHandler() to deal with received signal.
Before this commit, genericStreams() installed a signal handler for SIGINT and
SIGTERM.
Thus, these two signals were caught by both sigHandler() and genericStreams().
We suspected this to lead to race conditions where sigHandler() would be called,
then the trace would be deleted and completion would contain a message.
As a consequence, if failed some integration tests.

With this commit, sigHandler() is now the unique signal handler for the CLI.
It has the same goals as before but it now also handles printing
"\nTerminating..." when ^C is received while installed in
RunTraceAndPrintStream() (i.e. mainly trace gadgets) when output mode is not
JSON.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>