capabilities: add --unique flag
On some workloads, the capabilities trace gadget has a lot of repeated
output and it makes the gadget not easily usable.

There was already some BPF code to implement uniqueness with pid or
cgroup but it was not plugged to the CLI. This patch reuses the 'seen'
map that implemented the uniqueness logic but using the mntns id
instead.