pkg/runcfanotify: Use HOST_ROOT env variable
The runcfanotify package is designed to be used inside or outside a
container. When it's used inside a container it needs to access files
on the host, this is done in Inspektor Gadget by mounting the whole
host filesystem to /host. Let's update the logic on this package to use
the HOST_ROOT variable, if this is defined  then this prefix is added
to all paths to be secure the package can access them.

This commit fixes a weird issue where runcanotify wasn't detecing
containers because there was a bad path used:

error monitoring runc instance: cannot mark /var/run/foo: fanotify: mark error, no such file or directory