Introduce ConnectToContainerCollection()
ConnectToContainerCollection() is a helper function to connect
networking gadgets to the container collection. Networking gadgets
require to be attached to each container when they are created,
then we need the help of the container-collection package to get those
notifications.

The idea of this function is to be reusable among different networking
gadgets.