pkg/runcnotify: Use self-pipe trick to wake up poll()
The runcnotify package uses poll() to monitor the file descriptors (pidfd)
of containers to detect when they are removed. A 1000ms timeout is used
to make poll() return and update the list of file descriptors to monitor.
This solution is not perfect since after a creation of a container, this
mechanism could take up to 1 second to start monitoring for its deletion,
also, from a performance point of view to unblock each second to rebuild
the file descriptors list is not optimal.

This commit uses the self-pipe trick[0]: an end of the pipe is monitored
by poll() and the other end is written by another goroutine to unblock
it. This approach updates the list of watched fds faster as we don't have
to wait for the timeout, also this timeout is completed removed and poll()
only returns when there is an event on some file descriptors.

[0]: https://man7.org/linux/man-pages/man2/pselect.2.html