ci: Invalidate install-debian-packages cache when files are changed.
Before this commit, each time you changed the installed packages, you needed to
manually update the cache key.
With this commit, the cache key is now crafted as the hash of the
install-debian-packages action file and the main workflow file.
So, any modifications to these two files will invalidate the cache

Suggested-by: Burak Ok <burakok@microsoft.com>
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>