pkg/trace-collection: Avoid capturing events from wrong container
Inspektor Gadget uses the mnt namespace inode id of the containers to
implement the logic that determines if a given event produced in the
kernel should be traced or not.  There is a corner case when this logic
traces the events from the wrong container:

1. A tracer is filtering events from the "foo" container
2. The foo container is removed
3. A bar container is created
4. The tracer displays events from the "bar" container

This can happen because:
- The bar container can reuse the mnt ns inode id from the foo container
- The deletion of mount namespace is asynchronous, we can't block it
- Hence there is a window when the mnt ns inode id from the bar
  container is in the filtering map and hence events coming from it are
  wrongly captured.

This commit fixes it by keeping a file descriptor of the mnt namespace
of the container, guaranteeing that its inode id is not reused until
the tracer has updated all eBPF maps.