gadgets/profile: Fix bad stop logic
If there is an error calling tracer.Stop(), t.started will continue
to be always true, and further calls to Stop() will fail, hence the
tracer will enter a non recoverable state. This commit sets t.started
to false on Stop() whether there is an error or not. After this, the
error could be recovered by calling Start() again.