tools: Remove WORKING_SUFFIX from image-tag script.
In tools/image-tag, the output of git status --porcelain is analyzed to
determine if we should append "-WIP" to the image tag.
Sadly, grep 3.8 prints a warning message when running this regular
expression [1, 2].

This commit removes the existing regex and the logic around WORKING_SUFFIX.
Indeed, the previous regex was wrong as it tries to use non capturing group
('?:' operator), which are not available in grep with the extended regex engine.
As a consequence, the colon was treated as a character but this character is
never printed by git status --porcelain.
Then, the left part of the or operator (i.e. ?:[^?][^ ]) was always wrong and
only the right one could be right ([^ ][^?]).

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1] https://git.savannah.gnu.org/cgit/grep.git/commit/?h=v3.8&id=154661743525
[2] https://github.com/inspektor-gadget/inspektor-gadget/issues/1192#issue-1487457701