Include DNS response code in DNS trace events
The DNS response header RCODE field indicates the result of
a DNS query, allowing the user to distinguish successful queries
from queries that fail with NXDomain or other errors.

The local-gadget integration tests have been updated to validate
the new field for both success and NXDomain queries. This test
implicitly covers the changes to the DNS byte order in PR #1196
as well (without that change, the RCODE field would be FormErr
or Refused even for successful responses).