Add latency field to DNS trace gadget
Derive DNS latency from the request and response timestamps.
To avoid unbounded memory usage, set a limit on the number of
outstanding requests, dropping old requests to make space if necessary.

DNS requests are tracked in userspace. I tried using BPF_MAP_TYPE_LRU_HASH
in the eBPF program, but discovered that the LRU cache would evict
entries when the program was executed on different CPUs, even when
there were only a few entries in the map. For example, if nslookup
queried for A and AAAA records, the second request executing
on a different CPU could cause the first request to be evicted
from the map. The userspace implementation avoids this issue.

Signed-off-by: Will Daly <will.e.daly@gmail.com>