kubectl-gadget: Check kubernetes.io/os label only if --print-only was not given.
In commit b8fc17b561d2 ("kubectl-gadget: Only support Linux nodes explicitly"),
we added an os label to ensure Inspektor Gadget is only deployed on Linux node.
Sadly, this label is not available to all k8s version, so for older version we
remove this label.
But, to check the kubernetes version we need to communicate with remove cluster,
this leaded to error when running kubectl gadget deploy --print-only was not all
the YAML would be printed.

This commits removes the k8s version check is --print-only was given.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>