integration: Use /bin/sleep rather than sleep.
Busybox commit 58598eb70935 ("ash: optional sleep builtin") compiles by default
sleep as a builtin [1].
This commit is part of release 1.36.0 which was released 3 days ago [2].

As a consequence, TestExecsnoop which awaits an events with command sleep was
failing.
This commit forces using /bin/sleep rather than the builtin so our integration
test passes.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1] https://git.busybox.net/busybox/commit/?id=58598eb7093561d914a6254697e137b815f1fdfc
[2] https://git.busybox.net/busybox/commit/?id=70f77e4617e06077231b8b63c3fb3406d7f8865d