btfgen: Process one archive at a time.
Before this commit, we extracted all .tar.xz files and then call btfgen on them.
Sadly, this leaded to reaching the 14 GB disk quota of GitHub runner.

So, in this commit, we extract one archive, call btfgen on it and remove the
original archive and extracted file before processing another archive.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>