githubactions: Disable fail-fast
From documentation [0]:

"jobs.<job_id>.strategy.fail-fast applies to the entire matrix.
If jobs.<job_id>.strategy.fail-fast is set to true, GitHub will cancel
all in-progress and queued jobs in the matrix if any job in the matrix
fails. This property defaults to true."

"fast-failing" hides the result of other jobs on the same matrix, hence
we don't know if the issue is specific to that job or something more
generic. For instance, in [1] we not that the "Integration tests on AKS (Ubuntu, amd64)"
test failed, however we don't know the result of other OS / arch
combinations, and that information could help to have a better
understanding of what's going on.

[0]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstrategyfail-fast
[1]: https://github.com/inspektor-gadget/inspektor-gadget/actions/runs/4115766320/jobs/7105471579