trace network: don't hide host netns traffic
The trace network gadget currently doesn't capture traffic from pods
with hostNetwork set to true. This was intentional because this gadget
was initially designed for the needs of the network policy advisor.
Kubernetes network policies don't apply on the host network namespace,
so capturing the host network traffic was not necessary. However, the
trace network gadget has use cases outside of the network policy
advisor.