Dockerfiles: add socat to gadget container
By adding socat to the gadget container, we're able to exec into the pod and using
socat to connect to the unix socket to talk to the gadget-service. In the future we
could also expose the gadget-service using tcp and then either talk directly to it
or using the port-forward mechanism of kubernetes.