trace/network: Use ticker to query results
The trace network gadget is designed to give a summary of the network
connections performed by a pod / container. The eBPF program updates
a map that is then read by the user space part. This operation happens
each second, this is a tradeoff between reporting new connections fast
enough and not using too many CPU to poll the maps.

This mechanism was wrongly removed in
9a0941298c63 ("update gadgets to use new interfaces"), this commit
adds it again.

Fixes: 9a0941298c63 ("update gadgets to use new interfaces")