gadgets: use __always_inline
It seems that using "static inline" doesn't force the compiler to inline
the function in all cases, in some situation it caused the compiler
to emit a bpf-2-bpf call. This was creating issues for the trace sni
gadget as it was failing with: "invalid argument: BPF_LD_[ABS|IND] instructions cannot be mixed with bpf-to-bpf calls (1343 line(s) omitted)"

This commit changes it to use __always_inline in all gadgets, but there
is only effect on trace sni as the compiled object in the other ones is
the same.

Fixes: 8b01edb3d293 ("Use clang-15 to compile eBPF files")