pkg/gadgets: Use LINUX_KERNEL_VERSION for capabilities eBPF code.
When we first added this eBPF code, we were not able to compute in eBPF the
value of audit and insetid as computing these values relied on linux kernel
version.
So, we computed them in golang.

But, with the recent support of LINUX_KERNEL_VERSION added in cilium/ebpf, we
can now compute these values in eBPF code [1].
Note that, it makes our code looking more like upstream [2].

As a consequence, we can now remove the logic to compute insetID and audit from
golang code.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1] https://github.com/cilium/ebpf/commit/b6738990cfa3
[2] https://github.com/iovisor/bcc/blob/9371e844599f/libbpf-tools/capable.bpf.c#L122-L129