trace/tcpdrop: Don't use eventtypes.Event as base
eventtypes.Event has container related fields like namespace, pod,
container, however this gadget doesn't support it and using this
approach forces it to manually hide some columns and skip some parameters
that it doesn't support.

This commit includes directly the fields that are relevant for this
gadget. A cleaner solution to fix this for all gadgets should be
implemented, however it implies changing a lot of things so let's
go with a workaround for the time being.