kubeipresolver: refactor in preparation for use in tcpdrop
Currently, kubeipresolver is only used by the 'trace network' gadget and
it can resolve 1 local IP and 1 remote IP. In the 'trace tcpdrop'
gadget, we will need to resolve 2 IPs (source IP and destination IP) and
they might not be local IPs (i.e. not tied to a pod).

This refactoring does not change anything for the 'trace network' gadget
but will allow kubeipresolver to be used by other gadgets.