pkg/gadgets: Add static assert to ensure traceloop event struct have different sizes.
In the tracer, we use the size of object read from perf buffer to differentiate
between the two types of event (normal one and continued one).
So, they must have different size to be able to differentiate.

This commit then adds a static assert to check they have different size at
compile time.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>