pkg/gadget-collection: Store already detached containers in a map for traceloop.
The kernel can reuse mount namespace ID, so we need to beware of not detaching a
container which ID was reused by the kernel but what we do not care about.

So, before detaching a container, we check that its mount namespace ID was not
already detached by checking if it is a in map.
If this is not the case, we then detach it and add its ID to the map.
Otherwise, we just ignore it.

Suggested-by: Mauricio Vásquez <mauriciov@microsoft.com>
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>