Initial support for containerized gadgets
Containerized gadgets is a mechanism to run gadgets that are not builtin
within Inspektor Gadget. This support will allow in the future to
run gadgets developed by different parties. A good analogy is to that
Inspektor Gadget is for eBPF what docker is for containers.

This commit introduces some initial support for this. The supported
features are:
- Loading and attaching the different eBPF programs
- Polling the perf ring buffer and formatting the events based on their
  BTF information
- Enriching and filtering events by mount namespaces inode id

This support is experimental and will be extended / changed in follow
up PRs.

Some limitations are:
- Only tracer-like gadgets are supported

Signed-off-by: Burak Ok <burakok@microsoft.com>
Co-Authored-By: Mauricio Vásquez <mauriciov@microsoft.com>
Co-Authored-By: Alban Crequy <albancrequy@microsoft.com>