cri: Support CRI API v1/v1alpha2 simultaneously
containerd/crio will only support CRI API v1 in future releases so it
makes sense to support v1/v1alpha2 simultaneously. This change will make
v1 as the default version to talk to CRI and fallback to v1alpha2 in case
v1 is unimplemented. It is mainly based on the implementation in kubelet.