gadgets: Remove mntns filtering logic on exit probes
There are gadgets that install two probes on a given event: (1) one when
the thing is about to be executed and (2) another one when the thing is
done. The first one is used to collect some context information and the
second generates the event. In those cases, it's only needed to perform
the filtering logic on the first probe, the second one will do nothing
as the entry on the "start" (or similar) eBPF map is not created by the
first probe.

This also moves the filtering logic by mntns to entry probe for the
trace tcpconnect gadget.