cmd: exclude instead of include columns
Previously, we included columns specifically if they had a "kubernetes" tag for `kubectl-gadget`,
and if they had a "runtime" tag for "ig" - while this was the cleaner solution overall, it rendered
the column tags pretty much useless for anything else, because we also specifically added columns
without tags using `WithoutTags()`.
After this change we now exclude columns tagged with "kubernetes" for "ig" and columns
tagged with "runtime" for "kubectl-gadget". By preventing the use of `WithoutTags()`, we now can
use the tags for other purposes as well again.