ci: Mark release as draft.
Draft release are not published, so it permits us testing the release pipeline
without releasing something.
Moreover, we always edit the release notes once the release is done, so with
draft we can write the notes and publish it once satisfied with the notes.

Note that, the gadget kubectl plugin will be pushed to krew once the release is
made public thanks to the new release job.

Suggested-by: Jose Blanquicet <josebl@microsoft.com>
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>