runcfanotify: add missing /host for pid file
When ig or gadgettracermanager run inside a container, runcfanotify
monitors execution of runc on the host (e.g. /host/usr/bin/runc) and
then monitors the pid file such as:
/run/containerd/io.containerd.runtime.v2.task/moby/d29.../init.pid

The Go fanotify package provides GetPath() to gives the path of a file
descriptor of the pid file. But it provides the path from the host point
of view, without the /host part.

This is usually fine anyway because gadgettracermanager runs in a
container with a volume /run:/run, so it can find the pid file
correctly.

But in some conditions, it does not work: the Docker Desktop OS:
- uses a path like /var/run/desktop-containerd/daemon/io.containerd.runtime.v2.task/moby/1c5.../init.pid
- does not have the symlink /var/run->/run but the gadget container
  has it.

So when Inspektor Gadget tries to read /var/run/desk... in its
container, it follows its symlink /var/run->/run, then /run is the
volume. But host /run/desk... directory does not exist.

Symptoms:
```
WARN[0010] error watching pid: stat /var/run/desktop-containerd/daemon/io.containerd.runtime.v2.task/moby/1c5.../init.pid: no such file or directory
```

This patch access the pid file via the /host volume when configured, so
it does not rely on the /run:/run volume or the /var/run->/run symlink
on the host OS.