socket enricher: add socket earlier at tcp_connect
Before this patch, the first TCP-SYN packet could not be enriched
because the socket is added in the sockets map only at kretprobe-time
when tcp_v4_connect() and tcp_v6_connect() returns.

This patch changes this to update the sockets map at kprobe-time instead
of kretprobe. If the connection fails, the kretprobe detects it and
marks the socket for removal after the timeout.

Note that tcp_v4_connect() and tcp_v6_connect() both call tcp_connect(),
so the kprobe/kretprobe can be attached to tcp_connect directly.

The source port is updated to its final value by tcp_v4_connect() and
tcp_v6_connect() but before calling tcp_connect(), so attaching to
tcp_connect instead of the v4/v6 versions is necessary to get the
correct source port.