Add experimental feature
Inspektor Gadget is evolving very fast and we're implementing a lot of
new functionalities. Some of those functionalities are not fully ready
to be used, hence it's a good idea to disable them by default and only
enable them when they're mature enough.

This commit introduces a mechanism to enable experimental features. It
introduces a new experimental package that different components can
use to check if this is enabled without relying on the internal details.

The current implementation is based on an environmental variable for ig
and a CLI flag for kubectl-gadget

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>