pkg/columns: add support for dynamic fields
This change adds support for adding dynamic fields that could for example be part of
a byte slice. With that it's now possible to use arbitrary data instead of compile
time bound go structs.

Changes have also been made to the textcolumns formatter to use the newer and more
performant non-reflection based implementations of acessing fields.

Signed-off-by: Michael Friese <mfriese@microsoft.com>