pkg/gadgets/top/ebpf: introduce cpu usage
This patch introduce cpu usage column.

For those who want to compare ebpf prog overheads with other
processes, runtime is not intuitive. Thus, It would be better if we
have columns to show cpu usage like what we see in top.

Here we add two types of cpu usage, TotalCpuUsage and PerCpuUsage
1. TotalCpuUsage: The definition is similar to that in `top`. It is defined
as: The ebpf prog's share of the elapsed CPU time since the last update,
expressed as a percentage of total CPU time. Because the ebpf prog's runtime
is adding up runtime on each core in kernel, it's possible that the cpuusage
here is over 100%.
2. PerCpuUsage: PerCpuUsage is TotalCpuUsage divided by the number of
cpu cores. It may be  more clear and useful when comparing cpu usage on
different nodes with different cpu. This value should be less than 100%.

Signed-off-by: Kangjie Xu <kanxu@ebay.com>