container-collection: Keep net ns inode id reference
We're hitting, again, the issue where the events are enriched with the
wrong container data, this time in networking gadgets. This is happenning
because in some situations the container collection can keep the reference
to a container that is not present anymore in the kernel (yes, I know
containers aren't a kernel concept but you get the point). In this case,
the kernel can reuse the net ns inode in a new container, making the
enrichment process to use the wrong container.

The solution proposed here is the same as 4dedafc8565d ("Fix race condition on enrichers")
but this time for the net ns.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>