integration: Increase wait duration in TestPrometheus
TestPrometheus failed few times in the CI without expected counter value.
We believe it might be related to wait duration after we generate the events.
The idea of this change is to increase the sleep interval to ensure we get
enough time to capture all the events. Also, it enables debug logging
for prometheus pod.

Signed-off-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>