entrypoint: Do not fail if os-release info is not available.
The entry point script reads the os-release information from
/host/etc/os-release. This is used to print some debug messages and to
gather the BTF files from btfhub.  In some cases, this file is not
available at that location, for instance in NixOS. This makes the
installation fail and hence IG is unusable there.

This commit modifies the logic not to return a hard error in that case,
as this information is not fundmental and Inspektor Gadget can work
without it.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>