cmd/common/registry: fix `--help` for gadgets without parser, fix `--verbose`
There was a bug, that prevented gadgets that don't use a parser, from outputting help
when requested. This fixes that. This also fixes `--verbose` not working properly on
gadgets anymore by manually checking the flag at a later stage.

Reason for the changed behavior is that we delayed flag parsing, so that we can handle
them more dynamically and still have proper `--help` output.

Fixes: 8511d3b25dd7 ("gadgets: Add custom formatting options")

Signed-off-by: Michael Friese <mfriese@microsoft.com>