integration/run: New helper functions for testing containerized gadgets
We can't use the traditional functions for the containerized gadgets.
They were directly unmarshaling into the Event object which we don't
have with containerized gadgets.

Most of the columns are dynamically created and therefore the output
json is also mostly dynamic and depends solely on the eBPF program that
is getting run.

So we have to handle the JSON objects itself rather than comparing go
event structs

Signed-off-by: Burak Ok <burakok@microsoft.com>