container images: add descriptions and other labels
The list of standard labels and annotations are defined here:
https://github.com/opencontainers/image-spec/blob/main/annotations.md#pre-defined-annotation-keys

Note: this patch is only adding the LABEL in Dockerfiles. The annotation
in the index will be in a future patch.

When visiting different tags on the following container image pages,
this currently says:

> About this version: No description provided

https://github.com/inspektor-gadget/inspektor-gadget/pkgs/container/ig
https://github.com/orgs/inspektor-gadget/packages

According to the following documentation, it is recommended to use
labels except for the description of multi-arch images where an
annotation should be defined in the GitHub Action.

https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry#labelling-container-images

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>