gadgets/run: Support ebpf iterators
This commit introduces support for ebpf iterators in the run command.

As in the tracers (print_ map) case, BTF is a key component as it's used
to convert binary data to a human readable format. In this case, getting
the BTF information is not straightforward as it's not present on the
iterator program directly, instead, the gadget developers have to define
a dummy variable to guarantee the compiler generates this information for
us:

const struct myevent *gadget_iter_type __attribute__((unused));

Inspektor Gadget loads and attachs all iterators programs of type task
it found. When the gadget is run, it reads the results of the iterator
and uses the BTF information provided above to format the data.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>